+++ common/src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/entities/ImprovedThrownItemRenderer.java
@@ -1,0 +1,52 @@
+package net.mehvahdjukaar.supplementaries.client.renderers.entities;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+import com.mojang.math.Axis;
+import net.minecraft.client.renderer.MultiBufferSource;
+import net.minecraft.client.renderer.entity.EntityRenderer;
+import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.client.renderer.entity.ItemRenderer;
+import net.minecraft.client.renderer.texture.OverlayTexture;
+import net.minecraft.client.renderer.texture.TextureAtlas;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.projectile.ItemSupplier;
+import net.minecraft.world.item.ItemDisplayContext;
+
+public class ImprovedThrownItemRenderer<T extends Entity & ItemSupplier> extends EntityRenderer<T> {
+    public static final float MIN_CAMERA_DISTANCE_SQUARED = 7;
+    private final ItemRenderer itemRenderer;
+    private final float scale;
+
+    public ImprovedThrownItemRenderer(EntityRendererProvider.Context context, float scale) {
+        super(context);
+        this.itemRenderer = context.getItemRenderer();
+        this.scale = scale;
+    }
+
+    public ImprovedThrownItemRenderer(EntityRendererProvider.Context context) {
+        this(context, 1.0F);
+    }
+
+    // better centered and can render on first 2 ticks
+    @Override
+    public void render(T entity, float entityYaw, float partialTick, PoseStack poseStack, MultiBufferSource buffer, int packedLight) {
+        if (entity.tickCount < 2 && this.entityRenderDispatcher.camera.getPosition().distanceToSqr(entity.position()) < MIN_CAMERA_DISTANCE_SQUARED) {
+            return;
+        }
+        poseStack.pushPose();
+        poseStack.translate(0, entity.getBbHeight() / 2, 0);
+        poseStack.scale(this.scale * 0.5f, this.scale * 0.5f, this.scale * 0.5f);
+        poseStack.mulPose(this.entityRenderDispatcher.cameraOrientation());
+        poseStack.mulPose(Axis.YP.rotationDegrees(180.0F));
+        this.itemRenderer.renderStatic(entity.getItem(), ItemDisplayContext.NONE, packedLight, OverlayTexture.NO_OVERLAY, poseStack, buffer, entity.level(), entity.getId());
+        poseStack.popPose();
+        super.render(entity, entityYaw, partialTick, poseStack, buffer, packedLight);
+    }
+
+    @Override
+    public ResourceLocation getTextureLocation(T entity) {
+        return TextureAtlas.LOCATION_BLOCKS;
+    }
+}
+