+++ common/src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/entities/SlingshotProjectileRenderer.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.supplementaries.client.renderers.entities;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
 import net.mehvahdjukaar.supplementaries.common.entities.SlingshotProjectileEntity;
+import net.mehvahdjukaar.supplementaries.common.entities.ThrowableBrickEntity;
 import net.mehvahdjukaar.supplementaries.configs.ClientConfigs;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
@@ -16,6 +17,8 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.projectile.ItemSupplier;
 import net.minecraft.world.item.ItemDisplayContext;
+
+import static net.mehvahdjukaar.supplementaries.client.renderers.entities.ImprovedThrownItemRenderer.MIN_CAMERA_DISTANCE_SQUARED;
 
 public class SlingshotProjectileRenderer<T extends SlingshotProjectileEntity & ItemSupplier> extends EntityRenderer<T> {
     private final ItemRenderer itemRenderer;
@@ -33,27 +36,30 @@ public class SlingshotProjectileRenderer<T extends SlingshotProjectileEntity & I
 
     @Override
     public void render(T entity, float pEntityYaw, float partialTicks, PoseStack poseStack, MultiBufferSource buffer, int light) {
+        if (entity.tickCount < 2 && this.entityRenderDispatcher.camera.getPosition().distanceToSqr(entity.position()) < MIN_CAMERA_DISTANCE_SQUARED) {
+            return;
+        }
         //centers everything to hitbox y = 0 (rendered hitbox will be lowered)
         poseStack.translate(0, -entity.getBbHeight() / 2f, 0);
-        if (entity.tickCount >= 3 || (this.entityRenderDispatcher.camera.getEntity().distanceToSqr(entity) >= 12.25D)) {
-            poseStack.pushPose();
-            poseStack.translate(0, 0.25, 0);
 
-            poseStack.mulPose(Axis.YN.rotationDegrees(180 - Mth.rotLerp(partialTicks, entity.yRotO, entity.getYRot())));
-            poseStack.mulPose(Axis.ZN.rotationDegrees(Mth.rotLerp(partialTicks, entity.xRotO, entity.getXRot())));
+        poseStack.pushPose();
+        poseStack.translate(0, 0.25, 0);
 
-            float scale = (float)(double)ClientConfigs.Items.SLINGSHOT_PROJECTILE_SCALE.get();
-            poseStack.scale(scale, scale, scale);
+        poseStack.mulPose(Axis.YN.rotationDegrees(180 - Mth.rotLerp(partialTicks, entity.yRotO, entity.getYRot())));
+        poseStack.mulPose(Axis.ZN.rotationDegrees(Mth.rotLerp(partialTicks, entity.xRotO, entity.getXRot())));
 
-            this.itemRenderer.renderStatic(entity.getItem(), ItemDisplayContext.NONE, light, OverlayTexture.NO_OVERLAY,
-                    poseStack, buffer, entity.level(), 0);
-            poseStack.popPose();
-            super.render(entity, pEntityYaw, partialTicks, poseStack, buffer, light);
-        }
+        float scale = (float) (double) ClientConfigs.Items.SLINGSHOT_PROJECTILE_SCALE.get();
+        poseStack.scale(scale, scale, scale);
+
+        this.itemRenderer.renderStatic(entity.getItem(), ItemDisplayContext.NONE, light, OverlayTexture.NO_OVERLAY,
+                poseStack, buffer, entity.level(), 0);
+        poseStack.popPose();
+        super.render(entity, pEntityYaw, partialTicks, poseStack, buffer, light);
+
     }
 
     @Override
-    public ResourceLocation getTextureLocation(SlingshotProjectileEntity entity) {
+    public ResourceLocation getTextureLocation(T entity) {
         return TextureAtlas.LOCATION_BLOCKS;
     }
 }